fix: mail goals during the whole day, not just the morning
Goal mail jobs can easily take 4 hours to run and sometimes
fail. Restricting the allowable mail window to three morning hours
only means many goal mails will be missed.

Extend the window to the whole day so that mails are actually sent. It
will still prefer to send mail in the morning because that is the
first time it could encounter the goal and it will mark it as sent
afterwards.