build: upgrade libsass-python from 0.10 to 0.11
0.11 features the `sassc` command-line executable
which can be used in favor of the `sass` Python package.
This is desirable because it will be much easier to invoke
the CLI from scripts/assets/compile-sass.sh (the upcoming
replacement for pavelib/assets.py) than it would be to invoke a
Python function.

For more details, see
docs/decisions/0016-build-assets-without-python.rst

Part of (TODO link to issue)