fix: remove unused settings from edx-paltform related to learner record MFE
[APER-2240]

The monolith no longer needs to understand how to build URLs to the Learner Record MFE. The Credentials IDA has logic to determine if (and how) a request should be redirected to this MFE. This PR remove unused settings from edx-platform.

We have already removed the use of these settings in #31672.