temp: add some logging around extended profile update functionality
I am trying to persist/save data through an "extended profile field" in Stage but the data isn't saving. I have this working locally with devstack but don't understand why it's not working in Stage.

This PR adds some logging so I can do some debugging in Stage and try and pinpoint if/where the issue is occurring.