fix: inappropriate instructions for inline PII sharing consent dialog
This commit upgrades the version of the lti-consumer-xblock library from version 8.0.0 to version 8.0.1. Version 8.0.1 includes a fix to the copy in the LTI PII sharing propmpt dialog displayed before an inline LTI launch that shares PII.

Please see the CHANGELOG entry for this version for a full description of the change: https://github.com/openedx/xblock-lti-consumer/blob/master/CHANGELOG.rst#801---2023-02-03.