fix: allow content libraries to be created through Django admin
Previously, CMS Django admin would demand that at least one
LTIConfig be associated with the new library. This is not
necessary, but Django thought it was, because the `blank`
flag was not enabled on the ContentLibrary model's
`authored_lti_config` many-to-many field.

This should not affect the experimental blockstore-based
library authoring frontend, which already allows libraries
to be created without any authorized_lti_configs.

This will not affect the existing modulestore-based
library authoring frontend, since it doesn't even use
any of the models in question.

A migration is included with this commit, but it should NOT
change MySQL schema, since `blank` is a Django-level validation
flag.