refactor!: remove unused update-assets-dev.sh
The file was introduced as part of a the "decentralized
devstack" spike in order to make a working paver-free Dockerfile
in the edx-platform root.

edx-platform now has a working Dockerfile, and it does
not use this script.

The paver-free-assets initiative is being continued here:
https://github.com/openedx/edx-platform/issues/31798
It will result in a new, prod-ready script.