chore: disable forum digest subscription section by default (#31459)
Current State of the feature

Initially, this feature used the Notifier Django app but was deprecated
and removed from the configuration in favor of edx-notifications.
More info about this deprecation can be found in the GitHub issue (
https://github.com/openedx/build-test-release-wg/issues/22).

The edx-notifications app was not supported though, so its repository
was archived and moved to openedx-unsupported/edx-notifications.
Check the deprecation ticket (https://github.com/openedx/edx-notifications/issues/253)
and corresponding discussion (https://discuss.openedx.org/t/deprecation-removal-edx-notifications-repository/6748)
for details.

What remains in the edx-platform

The forum initial page contains the how-to info with the notification
preferences toggle.

This section is responsible for toggling the "notification_pref" user
preference (aka NOTIFICATION_PREF_KEY). I didn't find any usage of this
preference across the platform, so I'm suggesting hiding it by default
by setting the ENABLE_FORUM_DAILY_DIGEST setting to `False`.