fix: remove DJANGO_SETTINGS_MODULE env from Dockerfile
This variable is being removed from development target of Dockerfile because, we are using `EDX_PLATFORM_SETTINGS` env variable for picking settings file to use with Django management commands. When
this env variable is set to `lms.envs.devstack_docker`, with dev image being used with devstack, it picks that settings for running tests too as it has higher precedence than the settings in ini
file. So removing this, so the test settings can be picked from ini file.