Ensure samesite=none logged_in cookies are deleted
Fixes ENG-53
Patch delete_logged_in_cookies to bypass Django < 3.2.7 (Maple+) delete_cookie.
It relies on prefix of __SECURE to determine secure cookies, so we have to use
set_cookie with an expired date.
Note this will still work with Session cookie domain middleware.
Taken from https://discuss.overhang.io/t/logged-in-cookies-not-deleted-on-logout-over-https-not-reproducible-on-edx-org/1011/6