feat: Add script to regenerate JWKs for LMS asymmetric JWT signing
Roll keys used for devstack and tests. These keys have the full set of
optional precomputed parameters (p, q, dp, dq, qi) which makes them
suitable for use with PyJWT and some other libraries. (pyjwkest allowed
the use of a partial set of optional params, but other libraries throw.)

This is preparatory to rolling JWKs in stage and prod, which clears the way
for discontinuing use of pyjwkest.

Also:

- Remove signing JWK from CMS devstack, as per TODO