fix: remove DJANGO_SETTINGS_MODULE env from Dockerfile (#32013)
* fix: remove DJANGO_SETTINGS_MODULE env from Dockerfile

This variable is being removed from development target of Dockerfile because, we are using `EDX_PLATFORM_SETTINGS` env variable for picking settings file to use with Django management commands. When
this env variable is set to `lms.envs.devstack_docker`, with dev image being used with devstack, it picks that settings for running tests too as it has higher precedence than the settings in ini
file. So removing this, so the test settings can be picked from ini file.

* fix: move apt dependencies in base layer

Previously, git and build-essentials were being installed in builder-production layer so these depenedencies were not available in the dev images that are being used as LMS and Studion devstack
images. So moving these dependencies to base layer so it is pre-installed in devstack images.