feat: pre-fill location to Unit URL in Studio sidebar
Studio has a panel on the right side in the Unit view that shows authors
how to create a jump_to_id link to that Unit from HTML components that
are elsewhere in the course. However, that link was the literal text:

  "/jump_to_id/<location ID>"

The expectation was that people would then copy the location ID as well,
and construct the link manually. This commit is a minor tweak that just
displays the link address with the location ID pre-filled. Doing this
does cause a little bit of text overflow, which is why this commit also
removes the quotes surrounding the URL. It's still doesn't always fit
perfectly, but the increased usability is worth the small visual issue.