temp: disable source comments & maps for CI purposes
We want to compare the legacy and new outputs for SCSS compilation.
However, there is one silly difference between the two that we can
not reconcile:

* The legacy build (using Python `sass.compile`) adds source comments
  when in dev mode, but cannot output source maps.
* The new build (using CLI `sassc`) can only add source comments
  if it generates source maps.

So, for the purposes of making our comparison pass in CI, this commit:

* removes source_comments from legacy build, and
* removes --sourcemap from new build.

When we finally merge everything in, though, and no longer need this
comparison in CI, we will just enable source maps on the new build.