feat: Upgrade Python dependency lti-consumer-xblock
This commit upgrades the version of the lti-consumer-xblock library from version 9.0.4 to version 9.1.0. Version 9.1.0 adds support for sending a learner's full name in an LTI launch. The full name is sent as an LTI parameter in LTI 1.1 launches as the "lis_person_name_full" parameter. The full name is sent as an LTI parameter to LTI 1.3 launches as the "name" ID token claim.

Please see the CHANGELOG entry for these versions for a full description of the changes: https://github.com/openedx/xblock-lti-consumer/blob/master/CHANGELOG.rst#910---2023-04-28

This commit also downgrades the edx-proctoring-proctortrack dependency from version 1.2.1 to 1.0.5. This is due to a pin in the version of this requirement to 1.0.5 that was not removed when the requirement version was manually updated to version 1.2.1 in #32154. The newer version of the library does not contain any Python changes, so we are rolling back this version until further notice.