build: common/static/css/xmodule -> xmodule/static/css
Now that all XModule SCSS is located in xmodule/static/sass,
it would make sense to co-locate the CSS there as well.

We also add a README to explain the purpose of this new folder.

In the future, we will move xmodule/js and xmodule/assets
into xmodule/static as well.

Part of: https://github.com/openedx/edx-platform/issues/32292