feat: Pin python3-saml again for duplicate attributes issue (#32328)
We're seeing an issue with repeated attributes causing auth failure for
some IdPs, and need to take some time to investigate. See unpinning issue
https://github.com/openedx/edx-platform/issues/32327 for details.

This is expected to reintroduce some bugs in testing SAML in devstack.

This has the effect of reverting the following PRs:
- Unpinned: https://github.com/openedx/edx-platform/pull/32167 (partial)
- Upgraded: https://github.com/openedx/edx-platform/pull/32168

However, we're keeping the safe_lxml adjustments from the first PR, so it
should be fine on that front to re-upgrade whenever we're ready.