fix: allow only enrolled users in course forum roles
The course team management section under Instructor > Membership tab
allows users to be added a role even if are not enrolled in the course.
This is not the expected behaviour based on the help text displayed in
the section.

This PR updates update_forum_role_membership api to check whether user
is enrolled before adding them to a role.