fix: Add names to the error page urls.
Some of the static_template_view tests use names to get the URLs for the
error pages for testing, so I added names and updated the test to match
the new names.

I also updated the `test_404` function because we're no longer rendering
the 404 page in a different way from the 500 page so the response
includes the correct response code and content type.  This reduces the
number of differences between the 404 handler and the 500 handler.