temp: Ignore incorrect import for now
The commit for "feat: Allow delete course content in Studio only for admin users"
introduces an incorrect import of an lms package from cms. Fixing this requires
a larger refactoring that doesn't make sense as part of this rebase.

The upsteam version also has this issue at this time:
https://github.com/openedx/edx-platform/pull/27857

This can be dropped in Quince if the above is merged, or updated if the above
no longer has the issue.